Move test file directory to src/__tests__/
