+++ packages/akashic-cli-init/src/__tests__/BasicParametersSpec.ts
@@ -3,7 +3,7 @@ import * as path from "path";
 import {ConsoleLogger} from "@akashic/akashic-cli-commons/lib/ConsoleLogger";
 import * as fs from "fs-extra";
 import * as mockfs from "mock-fs";
-import * as bp  from "../lib/init/BasicParameters";
+import * as bp  from "../../lib/init/BasicParameters";
 import * as mockPrompt from "./support/mockPrompt";
 
 describe("BasicParameters", function () {
