+++ packages/akashic-cli-init/src/__tests__/initSpec.ts
@@ -2,8 +2,8 @@ import * as path from "path";
 import { ConsoleLogger } from "@akashic/akashic-cli-commons/lib/ConsoleLogger";
 import * as fs from "fs-extra";
 import * as mockfs from "mock-fs";
-import { internals } from "../lib/init/init";
-import { completeTemplateConfig } from "../lib/init/TemplateConfig";
+import { internals } from "../../lib/init/init";
+import { completeTemplateConfig } from "../../lib/init/TemplateConfig";
 const _extractFromTemplate =  internals._extractFromTemplate;
 
 describe("init.ts", () => {
