+++ packages/akashic-cli-init/src/__tests__/support/fixture/local/javascript-minimal/script/main.js
