+++ packages/akashic-cli-init/src/__tests__/support/fixture/remote/javascript.zip
