+++ packages/akashic-cli-init/src/__tests__/support/fixture/remote/typescript.zip
