+++ packages/akashic-cli-init/src/__tests__/support/mockConfigFile.ts
