+++ packages/akashic-cli-init/src/__tests__/support/mockPrompt.ts
